+++ test/acts_as_authentic_test/password_test.rb
@@ -51,7 +51,7 @@ module ActsAsAuthenticTest
     end
 
     def test_validates_length_of_password_field_options_config
-      default = {:minimum => 4, :if => :require_password?}
+      default = {:minimum => 8, :if => :require_password?}
       assert_equal default, User.validates_length_of_password_field_options
       assert_equal default, Employee.validates_length_of_password_field_options
 
@@ -73,7 +73,7 @@ module ActsAsAuthenticTest
     end
 
     def test_validates_length_of_password_confirmation_field_options_config
-      default = {:minimum => 4, :if => :require_password?}
+      default = {:minimum => 8, :if => :require_password?}
       assert_equal default, User.validates_length_of_password_confirmation_field_options
       assert_equal default, Employee.validates_length_of_password_confirmation_field_options
 
@@ -104,21 +104,21 @@ module ActsAsAuthenticTest
     end
 
     def test_validates_length_of_password
-      u = User.new(login: "abcde", email: "abcde@test.com", password: "abcde", password_confirmation: "abcde")
+      u = User.new(login: "abcde", email: "abcde@test.com", password: "abcdefgh", password_confirmation: "abcdefgh")
       assert u.valid?
 
-      u.password = u.password_confirmation = "abc"
+      u.password = u.password_confirmation = "abcdef"
       assert !u.valid?
 
-      assert u.errors[:password].include?("is too short (minimum is 4 characters)")
-      assert u.errors[:password_confirmation].include?("is too short (minimum is 4 characters)")
+      assert u.errors[:password].include?("is too short (minimum is 8 characters)")
+      assert u.errors[:password_confirmation].include?("is too short (minimum is 8 characters)")
     end
 
     def test_validates_confirmation_of_password
-      u = User.new(login: "abcde", email: "abcde@test.com", password: "abcde", password_confirmation: "abcde")
+      u = User.new(login: "abcde", email: "abcde@test.com", password: "abcdefgh", password_confirmation: "abcdefgh")
       assert u.valid?
 
-      u.password_confirmation = "abcdefgh"
+      u.password_confirmation = "abcdefghij"
       assert !u.valid?
 
       if ActiveModel.respond_to?(:version) and ActiveModel.version.segments.first >= 4
@@ -131,23 +131,23 @@ module ActsAsAuthenticTest
     def test_validates_length_of_password_confirmation
       u = User.new
 
-      u.password = "test"
+      u.password = "testpass"
       u.password_confirmation = ""
       assert !u.valid?
       assert u.errors[:password_confirmation].size > 0
 
-      u.password_confirmation = "test"
+      u.password_confirmation = "testpass"
       assert !u.valid?
       assert u.errors[:password_confirmation].size == 0
 
       ben = users(:ben)
       assert ben.valid?
 
-      ben.password = "newpass"
+      ben.password = "newpasswd"
       assert !ben.valid?
       assert ben.errors[:password_confirmation].size > 0
 
-      ben.password_confirmation = "newpass"
+      ben.password_confirmation = "newpasswd"
       assert ben.valid?
     end